+++ src/app/(management)/dashboard/DashboardProvider.tsx
@@ -1,13 +1,16 @@
 import { AppProvider } from "@/contexts/AppContext"
 import { AuthProvider } from "@/contexts/JWTContext"
 import DashBoardComponent from "./DashBoardComponent"
+import AuthGuard from "@/guards/AuthGuard"
 
 
-const Dashboard = (props : {}) => {
+const Dashboard = (props: {}) => {
     return (
         <AppProvider>
             <AuthProvider>
-                <DashBoardComponent/>
+                <AuthGuard>
+                    <DashBoardComponent />
+                </AuthGuard>
             </AuthProvider>
         </AppProvider>
     )
