+++ src/app/(management)/dashboard/page.tsx
@@ -1,17 +1,17 @@
 import { Metadata } from "next";
-import OrderPageProvider from "./components/OrderPageProvider";
 import MainLayout from "@/components/layouts";
+import Dashboard from "./DashboardProvider";
 
 const metadata: Metadata = {
   title: "Loco. - Quản Lý",
   description: "This is Loco. Category Management Page",
 };
 
-export default function OrderManagementPage() {
+export default function DashboardManagementPage() {
   return (
     <>
       <MainLayout metadata={metadata}>
-        <OrderPageProvider />
+        <Dashboard/>
       </MainLayout>
     </>
   );
