+++ src/app/page.tsx
@@ -9,7 +9,7 @@ export const metadata: Metadata = {
 export default function Home() {
   return (
     <>
-      <HomePageProvider />
+      <HomePageProvider metadata={metadata} />
     </>
   );
 }
