+++ src/app/pageComponent.tsx
@@ -1,16 +1,14 @@
-"use client";
+// "use client";
 
-import Dashboard from "@/components/layouts/dashboard";
-import HomePageProvider from "./pageProvider";
+// import Dashboard from "@/components/layouts/dashboard";
+// import HomePageProvider from "./pageProvider";
 
-const HomePageComponent = (props: {}) => {
-  return (
-    <>
-      <HomePageProvider />
-      <div style={{  }}>
-      </div>
-    </>
-  );
-};
+// const HomePageComponent = () => {
+//   return (
+//     <>
+//       <HomePageProvider />
+//     </>
+//   );
+// };
 
-export default HomePageComponent;
+// export default HomePageComponent;
