+++ src/components/layouts/category/index.tsx
@@ -1,16 +1,16 @@
-import { AppProvider } from "@/contexts/AppContext";
-import { AuthProvider } from "@/contexts/JWTContext";
-import CategoryComponent from "./CategoryComponent";
-import MainLayoutComponent from "../layout";
+// import { AppProvider } from "@/contexts/AppContext";
+// import { AuthProvider } from "@/contexts/JWTContext";
+// import CategoryComponent from "./CategoryComponent";
+// import MainLayoutComponent from "../layout";
 
-const Category = (props: {}) => {
-  return (
-    <AppProvider>
-      <AuthProvider>
-        <CategoryComponent/>
-      </AuthProvider>
-    </AppProvider>
-  );
-};
+// const Category = (props: {}) => {
+//   return (
+//     <AppProvider>
+//       <AuthProvider>
+//         <CategoryComponent/>
+//       </AuthProvider>
+//     </AppProvider>
+//   );
+// };
 
-export default Category;
+// export default Category;
