+++ src/components/layouts/index.tsx
@@ -1,10 +1,12 @@
 import React from "react";
 import MainLayoutComponent from "./layout";
+import { Metadata } from "next";
 
 export default function MainLayout({
-  children,
+  children, metadata
 }: {
   children: React.ReactNode;
+  metadata: Metadata
 }) {
   return (
     <>
