+++ src/components/layouts/order/OrderComponent.tsx
@@ -1,142 +1,142 @@
-import React from "react";
-import { OrderTable } from "../table/OrderTable";
+// import React from "react";
+// import { OrderTable } from "../table/OrderTable";
 
-const OrderComponent: React.FC = () => {
-  return (
-    <div>
-      <div
-        className="relative flex min-h-screen flex-col items-center overflow-hidden"
-        style={{ backgroundColor: "#F1F5F9" }}
-      >
-        <div className="relative mx-auto">
-          <div className="group relative flex cursor-pointer after:shadow-lg after:shadow-black">
-            {/* hidden slide */}
-            <div
-              className="relative -left-16 top-0 z-10 w-96 rounded-xl px-5 py-3 text-base font-semibold leading-7 transition-all duration-700 group-hover:-left-14"
-              style={{ backgroundColor: "#8a9db5" }}
-            >
-              <div className="flex flex-col gap-4">
-                {/* item 1 */}
-                <div className="flex items-center gap-2">
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    width="20px"
-                    height="20px"
-                    viewBox="0 0 24 24"
-                    fill="none"
-                  >
-                    <path
-                      fill-rule="evenodd"
-                      clip-rule="evenodd"
-                      d="M4.63338 2.58607C4.63338 1.87578 5.20918 1.29999 5.91946 1.29999H18.0806C18.7869 1.29999 19.3667 1.86927 19.3667 2.58352V2.59367V2.60383V2.614V2.62418V2.63437V2.64458V2.65479V2.66502V2.67526V2.68551V2.69577V2.70604V2.71633V2.72662V2.73693V2.74725V2.75758V2.76792V2.77827V2.78863V2.799V2.80939V2.81978V2.83019V2.8406V2.85103V2.86147V2.87192V2.88238V2.89285V2.90333V2.91382V2.92433V2.93484V2.94536V2.9559V2.96644V2.977V2.98757V2.99814V3.00873V3.01933V3.02994V3.04056V3.05119V3.06183V3.07247V3.08313V3.09381V3.10449V3.11518V3.12588V3.13659V3.14731V3.15804V3.16878V3.17953V3.19029V3.20106V3.21184V3.22264V3.23344V3.24425V3.25507V3.2659V3.27674V3.28759V3.29845V3.30931V3.32019V3.33108V3.34198V3.35289V3.3638V3.37473V3.38567V3.39661V3.40757V3.41853V3.4295V3.44049V3.45148V3.46248V3.47349V3.48451V3.49554V3.50658V3.51763V3.52222H19.7778C21.3917 3.52222 22.7001 4.83054 22.7001 6.44444V7.89663C22.7001 9.023 22.0527 10.0491 21.0361 10.5341L18.9681 11.5207C18.903 11.5518 18.8354 11.5719 18.7675 11.5818C17.726 13.9966 15.4297 15.7432 12.7001 16.0005V17.9667H14.5927C15.2249 17.9667 15.8402 18.1717 16.346 18.5511L17.5312 19.44C18.9236 20.4843 18.185 22.7 16.4445 22.7H7.55562C5.81513 22.7 5.07654 20.4843 6.46895 19.44L7.65413 18.5511C8.15996 18.1717 8.77519 17.9667 9.40747 17.9667H11.3001V16.0005C8.57046 15.7432 6.27425 13.9965 5.23266 11.5818C5.16475 11.5719 5.09716 11.5518 5.03197 11.5207L2.964 10.5341C1.9474 10.0491 1.30005 9.023 1.30005 7.89663V6.44444C1.30005 4.83054 2.60837 3.52222 4.22227 3.52222H4.63338V2.58607ZM17.9667 8.66687C17.9667 11.9614 15.2966 14.6321 12.0023 14.6333H12C8.70475 14.6333 6.03338 11.962 6.03338 8.66665V2.69999H17.9667V2.70604V2.71633V2.72662V2.73693V2.74725V2.75758V2.76792V2.77827V2.78863V2.799V2.80939V2.81978V2.83019V2.8406V2.85103V2.86147V2.87192V2.88238V2.89285V2.90333V2.91382V2.92433V2.93484V2.94536V2.9559V2.96644V2.977V2.98757V2.99814V3.00873V3.01933V3.02994V3.04056V3.05119V3.06183V3.07247V3.08313V3.09381V3.10449V3.11518V3.12588V3.13659V3.14731V3.15804V3.16878V3.17953V3.19029V3.20106V3.21184V3.22264V3.23344V3.24425V3.25507V3.2659V3.27674V3.28759V3.29845V3.30931V3.32019V3.33108V3.34198V3.35289V3.3638V3.37473V3.38567V3.39661V3.40757V3.41853V3.4295V3.44049V3.45148V3.46248V3.47349V3.48451V3.49554V3.50658V3.51763V3.52868V3.53975V3.55082V3.56191V3.573V3.5841V3.59521V3.60633V3.61746V3.6286V3.63975V3.6509V3.66206V3.67324V3.68442V3.69561V3.70681V3.71801V3.72923V3.74045V3.75169V3.76293V3.77418V3.78544V3.7967V3.80798V3.81926V3.83056V3.84186V3.85317V3.86448V3.87581V3.88714V3.89848V3.90983V3.92119V3.93256V3.94393V3.95531V3.9667V3.9781V3.98951V4.00092V4.01234V4.02377V4.03521V4.04666V4.05811V4.06957V4.08104V4.09252V4.104V4.11549V4.12699V4.1385V4.15002V4.16154V4.17307V4.18461V4.19615V4.2077V4.21926V4.23083V4.2424V4.25398V4.26557V4.27717V4.28877V4.30038V4.312V4.32363V4.33526V4.3469V4.35854V4.3702V4.38186V4.39353V4.4052V4.41688V4.42857V4.44026V4.45197V4.46368V4.47539V4.48711V4.49884V4.51058V4.52232V4.53407V4.54582V4.55759V4.56936V4.58113V4.59291V4.6047V4.6165V4.6283V4.6401V4.65192V4.66374V4.67556V4.6874V4.69924V4.71108V4.72293V4.73479V4.74665V4.75852V4.7704V4.78228V4.79417V4.80607V4.81797V4.82987V4.84178V4.8537V4.86563V4.87755V4.88949V4.90143V4.91338V4.92533V4.93729V4.94925V4.96122V4.9732V4.98518V4.99717V5.00916V5.02115V5.03316V5.04517V5.05718V5.0692V5.08122V5.09325V5.10529V5.11733V5.12937V5.14142V5.15348V5.16554V5.17761V5.18968V5.20176V5.21384V5.22592V5.23802V5.25011V5.26221V5.27432V5.28643V5.29855V5.31067V5.32279V5.33492V5.34706V5.3592V5.37134V5.38349V5.39565V5.40781V5.41997V5.43214V5.44431V5.45649V5.46867V5.48085V5.49304V5.50524V5.51744V5.52964V5.54185V5.55406V5.56627V5.57849V5.59072V5.60295V5.61518V5.62742V5.63966V5.6519V5.66415V5.6764V5.68866V5.70092V5.71318V5.72545V5.73772V5.75V5.76228V5.77456V5.78685V5.79914V5.81144V5.82373V5.83604V5.84834V5.86065V5.87296V5.88528V5.8976V5.90992V5.92225V5.93458V5.94691V5.95925V5.97159V5.98393V5.99628V6.00863V6.02098V6.03334V6.0457V6.05806V6.07042V6.08279V6.09516V6.10754V6.11992V6.1323V6.14468V6.15707V6.16945V6.18185V6.19424V6.20664V6.21904V6.23144V6.24385V6.25626V6.26867V6.28108V6.2935V6.30591V6.31834V6.33076V6.34319V6.35561V6.36805V6.38048V6.39292V6.40535V6.41779V6.43024V6.44268V6.45513V6.46758V6.48003V6.49248V6.50494V6.5174V6.52986V6.54232V6.55478V6.56725V6.57972V6.59219V6.60466V6.61713V6.62961V6.64208V6.65456V6.66704V6.67953V6.69201V6.7045V6.71698V6.72947V6.74196V6.75445V6.76695V6.77944V6.79194V6.80444V6.81694V6.82944V6.84194V6.85444V6.86695V6.87945V6.89196V6.90447V6.91698V6.92949V6.942V6.95451V6.96703V6.97954V6.99206V7.00458V7.01709V7.02961V7.04213V7.05465V7.06717V7.0797V7.09222V7.10474V7.11727V7.12979V7.14232V7.15485V7.16737V7.1799V7.19243V7.20496V7.21749V7.23002V7.24255V7.25508V7.26761V7.28014V7.29268V7.30521V7.31774V7.33027V7.3428V7.35534V7.36787V7.38041V7.39294V7.40547V7.41801V7.43054V7.44307V7.45561V7.46814V7.48068V7.49321V7.50574V7.51828V7.53081V7.54334V7.55587V7.56841V7.58094V7.59347V7.606V7.61853V7.63106V7.64359V7.65612V7.66865V7.68118V7.69371V7.70624V7.71876V7.73129V7.74381V7.75634V7.76886V7.78139V7.79391V7.80643V7.81895V7.83147V7.84399V7.85651V7.86903V7.88154V7.89406V7.90657V7.91909V7.9316V7.94411V7.95662V7.96913V7.98163V7.99414V8.00665V8.01915V8.03165V8.04415V8.05665V8.06915V8.08165V8.09414V8.10664V8.11913V8.13162V8.14411V8.1566V8.16908V8.18157V8.19405V8.20653V8.21901V8.23149V8.24396V8.25644V8.26891V8.28138V8.29385V8.30631V8.31878V8.33124V8.3437V8.35616V8.36862V8.38107V8.39352V8.40597V8.41842V8.43087V8.44331V8.45575V8.46819V8.48063V8.49306V8.50549V8.51792V8.53035V8.54277V8.55519V8.56761V8.58003V8.59244V8.60486V8.61726V8.62967V8.64207V8.65448V8.66687ZM19.2766 9.82233L20.4333 9.27052C20.9628 9.01788 21.3001 8.48337 21.3001 7.89663V6.44444C21.3001 5.60374 20.6185 4.92222 19.7778 4.92222H19.3667V4.92533V4.93729V4.94925V4.96122V4.9732V4.98518V4.99717V5.00916V5.02115V5.03316V5.04517V5.05718V5.0692V5.08122V5.09325V5.10529V5.11733V5.12937V5.14142V5.15348V5.16554V5.17761V5.18968V5.20176V5.21384V5.22592V5.23802V5.25011V5.26221V5.27432V5.28643V5.29855V5.31067V5.32279V5.33492V5.34706V5.3592V5.37134V5.38349V5.39565V5.40781V5.41997V5.43214V5.44431V5.45649V5.46867V5.48085V5.49304V5.50524V5.51744V5.52964V5.54185V5.55406V5.56627V5.57849V5.59072V5.60295V5.61518V5.62742V5.63966V5.6519V5.66415V5.6764V5.68866V5.70092V5.71318V5.72545V5.73772V5.75V5.76228V5.77456V5.78685V5.79914V5.81144V5.82373V5.83604V5.84834V5.86065V5.87296V5.88528V5.8976V5.90992V5.92225V5.93458V5.94691V5.95925V5.97159V5.98393V5.99628V6.00863V6.02098V6.03334V6.0457V6.05806V6.07042V6.08279V6.09516V6.10754V6.11992V6.1323V6.14468V6.15707V6.16945V6.18185V6.19424V6.20664V6.21904V6.23144V6.24385V6.25626V6.26867V6.28108V6.2935V6.30591V6.31834V6.33076V6.34319V6.35561V6.36805V6.38048V6.39292V6.40535V6.41779V6.43024V6.44268V6.45513V6.46758V6.48003V6.49248V6.50494V6.5174V6.52986V6.54232V6.55478V6.56725V6.57972V6.59219V6.60466V6.61713V6.62961V6.64208V6.65456V6.66704V6.67953V6.69201V6.7045V6.71698V6.72947V6.74196V6.75445V6.76695V6.77944V6.79194V6.80444V6.81694V6.82944V6.84194V6.85444V6.86695V6.87945V6.89196V6.90447V6.91698V6.92949V6.942V6.95451V6.96703V6.97954V6.99206V7.00458V7.01709V7.02961V7.04213V7.05465V7.06717V7.0797V7.09222V7.10474V7.11727V7.12979V7.14232V7.15485V7.16737V7.1799V7.19243V7.20496V7.21749V7.23002V7.24255V7.25508V7.26761V7.28014V7.29268V7.30521V7.31774V7.33027V7.3428V7.35534V7.36787V7.38041V7.39294V7.40547V7.41801V7.43054V7.44307V7.45561V7.46814V7.48068V7.49321V7.50574V7.51828V7.53081V7.54334V7.55587V7.56841V7.58094V7.59347V7.606V7.61853V7.63106V7.64359V7.65612V7.66865V7.68118V7.69371V7.70624V7.71876V7.73129V7.74381V7.75634V7.76886V7.78139V7.79391V7.80643V7.81895V7.83147V7.84399V7.85651V7.86903V7.88154V7.89406V7.90657V7.91909V7.9316V7.94411V7.95662V7.96913V7.98163V7.99414V8.00665V8.01915V8.03165V8.04415V8.05665V8.06915V8.08165V8.09414V8.10664V8.11913V8.13162V8.14411V8.1566V8.16908V8.18157V8.19405V8.20653V8.21901V8.23149V8.24396V8.25644V8.26891V8.28138V8.29385V8.30631V8.31878V8.33124V8.3437V8.35616V8.36862V8.38107V8.39352V8.40597V8.41842V8.43087V8.44331V8.45575V8.46819V8.48063V8.49306V8.50549V8.51792V8.53035V8.54277V8.55519V8.56761V8.58003V8.59244V8.60486V8.61726V8.62967V8.64207V8.65448V8.66687C19.3667 9.05998 19.3359 9.44589 19.2766 9.82233ZM4.63338 8.66665C4.63338 9.05984 4.66419 9.44583 4.72351 9.82235L3.56682 9.27052C3.03727 9.01788 2.70005 8.48337 2.70005 7.89663V6.44444C2.70005 5.60374 3.38157 4.92222 4.22227 4.92222H4.63338V8.66665ZM8.49413 19.6711C8.75763 19.4735 9.0781 19.3667 9.40747 19.3667H12.0001H14.5927C14.922 19.3667 15.2425 19.4735 15.506 19.6711L16.6912 20.56C17.0072 20.797 16.8396 21.3 16.4445 21.3H7.55562C7.16053 21.3 6.99289 20.797 7.30895 20.56L8.49413 19.6711Z"
-                      fill="white"
-                    />
-                    <path
-                      d="M11.8425 6.20125C11.9226 6.09893 12.0774 6.09893 12.1575 6.20125L12.7183 6.91782C12.7413 6.94724 12.7721 6.96963 12.8072 6.98244L13.662 7.29436C13.7841 7.3389 13.832 7.48617 13.7594 7.59395L13.2512 8.34874C13.2303 8.37973 13.2185 8.41595 13.2172 8.45329L13.1847 9.36265C13.1801 9.49249 13.0548 9.58351 12.9299 9.5478L12.055 9.29771C12.019 9.28744 11.981 9.28744 11.945 9.29771L11.0701 9.5478C10.9452 9.58351 10.8199 9.49249 10.8153 9.36265L10.7828 8.45329C10.7815 8.41595 10.7697 8.37973 10.7488 8.34874L10.2406 7.59395C10.168 7.48617 10.2159 7.3389 10.338 7.29436L11.1928 6.98244C11.2279 6.96963 11.2587 6.94724 11.2817 6.91782L11.8425 6.20125Z"
-                      fill="#DF1463"
-                    />
-                  </svg>
-                  <div className="flex items-center gap-1 rounded-full bg-green-400/25 py-0.5 pl-1 pr-2">
-                    <p className="-rotate-90 leading-tight text-green-500">
-                      &rarr;
-                    </p>
-                    <p className="text-xs leading-tight text-white">4%</p>
-                  </div>
-                  <p className="text-white opacity-0 delay-200 duration-700 ease-in-out group-hover:opacity-100">
-                    75 Orders
-                  </p>
-                </div>
-                {/* item 2 */}
-                <div className="flex items-center gap-2">
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    width="20px"
-                    height="20px"
-                    viewBox="0 0 24 24"
-                    fill="none"
-                  >
-                    <path
-                      d="M3.7644 13.3726V11.0849C3.7644 6.53663 7.45154 2.84949 11.9999 2.84949V2.84949C16.5482 2.84949 20.2353 6.53663 20.2353 11.0849V13.9856"
-                      stroke="#DF1463"
-                      strokeWidth="1.69904"
-                    />
-                    <path
-                      d="M2.84937 17.4903V16.1566C2.84937 14.1351 4.4881 12.4964 6.50957 12.4964C7.5203 12.4964 8.33967 13.3158 8.33967 14.3265V19.3204C8.33967 20.3311 7.5203 21.1505 6.50957 21.1505C4.4881 21.1505 2.84937 19.5118 2.84937 17.4903Z"
-                      stroke="white"
-                      strokeWidth="1.69904"
-                    />
-                    <path
-                      d="M21.1504 17.4903V16.1566C21.1504 14.1351 19.5117 12.4964 17.4902 12.4964C16.4795 12.4964 15.6601 13.3158 15.6601 14.3265V19.3204C15.6601 20.3311 16.4795 21.1505 17.4902 21.1505C19.5117 21.1505 21.1504 19.5118 21.1504 17.4903Z"
-                      stroke="white"
-                      strokeWidth="1.69904"
-                    />
-                  </svg>
-                  <div className="flex items-center gap-1 rounded-full bg-green-400/25 py-0.5 pl-1 pr-2">
-                    <p className="-rotate-90 leading-tight text-green-500">
-                      &rarr;
-                    </p>
-                    <p className="text-xs leading-tight text-white">4%</p>
-                  </div>
-                  <p className="text-white opacity-0 delay-200 duration-700 ease-in-out group-hover:opacity-100">
-                    357 Delivered
-                  </p>
-                </div>
-                {/* item 3 */}
-                <div className="flex items-center gap-2">
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    width="20px"
-                    height="20px"
-                    viewBox="0 0 24 24"
-                    fill="none"
-                  >
-                    <path
-                      d="M3.7644 13.3726V11.0849C3.7644 6.53663 7.45154 2.84949 11.9999 2.84949V2.84949C16.5482 2.84949 20.2353 6.53663 20.2353 11.0849V13.9856"
-                      stroke="#DF1463"
-                      strokeWidth="1.69904"
-                    />
-                    <path
-                      d="M2.84937 17.4903V16.1566C2.84937 14.1351 4.4881 12.4964 6.50957 12.4964C7.5203 12.4964 8.33967 13.3158 8.33967 14.3265V19.3204C8.33967 20.3311 7.5203 21.1505 6.50957 21.1505C4.4881 21.1505 2.84937 19.5118 2.84937 17.4903Z"
-                      stroke="white"
-                      strokeWidth="1.69904"
-                    />
-                    <path
-                      d="M21.1504 17.4903V16.1566C21.1504 14.1351 19.5117 12.4964 17.4902 12.4964C16.4795 12.4964 15.6601 13.3158 15.6601 14.3265V19.3204C15.6601 20.3311 16.4795 21.1505 17.4902 21.1505C19.5117 21.1505 21.1504 19.5118 21.1504 17.4903Z"
-                      stroke="white"
-                      strokeWidth="1.69904"
-                    />
-                  </svg>
-                  <div className="flex items-center gap-1 rounded-full bg-red-400/25 py-0.5 pl-1 pr-2">
-                    <p className="-rotate-n90 leading-tight text-red-500">
-                      &rarr;
-                    </p>
-                    <p className="text-xs leading-tight text-white">25%</p>
-                  </div>
-                  <p className="text-white opacity-0 delay-200 duration-700 ease-in-out group-hover:opacity-100">
-                    65 Canceled
-                  </p>
-                </div>
-              </div>
-            </div>
-            {/* main */}
-            <div
-              className="absolute -right-16 top-0 z-20 flex w-96 flex-col gap-4 self-end rounded-xl rounded-l-2xl border-none px-5 py-3 text-base font-semibold leading-7 transition-all duration-700 group-hover:-right-14 group-hover:w-64 group-hover:rounded-l-lg"
-              style={{ backgroundColor: "#B0C4DE" }}
-            >
-              <p className="text-[#fff]">Total Orders</p>
-              <p className="text-[#fff]">Total Delivered</p>
-              <p className="text-[#fff]">Total Canceled</p>
-            </div>
-            <div
-              className="h-16 w-[28rem] -left-10 shadow-2xl shadow-[#3d348b] absolute bottom-0"
-              style={{ backgroundColor: "#99CCFF" }}
-            ></div>
-          </div>
-        </div>
-        <OrderTable />
-      </div>
-    </div>
-  );
-};
+// const OrderComponent: React.FC = () => {
+//   return (
+//     <div>
+//       <div
+//         className="relative flex min-h-screen flex-col items-center overflow-hidden"
+//         style={{ backgroundColor: "#F1F5F9" }}
+//       >
+//         <div className="relative mx-auto">
+//           <div className="group relative flex cursor-pointer after:shadow-lg after:shadow-black">
+//             {/* hidden slide */}
+//             <div
+//               className="relative -left-16 top-0 z-10 w-96 rounded-xl px-5 py-3 text-base font-semibold leading-7 transition-all duration-700 group-hover:-left-14"
+//               style={{ backgroundColor: "#8a9db5" }}
+//             >
+//               <div className="flex flex-col gap-4">
+//                 {/* item 1 */}
+//                 <div className="flex items-center gap-2">
+//                   <svg
+//                     xmlns="http://www.w3.org/2000/svg"
+//                     width="20px"
+//                     height="20px"
+//                     viewBox="0 0 24 24"
+//                     fill="none"
+//                   >
+//                     <path
+//                       fill-rule="evenodd"
+//                       clip-rule="evenodd"
+//                       d="M4.63338 2.58607C4.63338 1.87578 5.20918 1.29999 5.91946 1.29999H18.0806C18.7869 1.29999 19.3667 1.86927 19.3667 2.58352V2.59367V2.60383V2.614V2.62418V2.63437V2.64458V2.65479V2.66502V2.67526V2.68551V2.69577V2.70604V2.71633V2.72662V2.73693V2.74725V2.75758V2.76792V2.77827V2.78863V2.799V2.80939V2.81978V2.83019V2.8406V2.85103V2.86147V2.87192V2.88238V2.89285V2.90333V2.91382V2.92433V2.93484V2.94536V2.9559V2.96644V2.977V2.98757V2.99814V3.00873V3.01933V3.02994V3.04056V3.05119V3.06183V3.07247V3.08313V3.09381V3.10449V3.11518V3.12588V3.13659V3.14731V3.15804V3.16878V3.17953V3.19029V3.20106V3.21184V3.22264V3.23344V3.24425V3.25507V3.2659V3.27674V3.28759V3.29845V3.30931V3.32019V3.33108V3.34198V3.35289V3.3638V3.37473V3.38567V3.39661V3.40757V3.41853V3.4295V3.44049V3.45148V3.46248V3.47349V3.48451V3.49554V3.50658V3.51763V3.52222H19.7778C21.3917 3.52222 22.7001 4.83054 22.7001 6.44444V7.89663C22.7001 9.023 22.0527 10.0491 21.0361 10.5341L18.9681 11.5207C18.903 11.5518 18.8354 11.5719 18.7675 11.5818C17.726 13.9966 15.4297 15.7432 12.7001 16.0005V17.9667H14.5927C15.2249 17.9667 15.8402 18.1717 16.346 18.5511L17.5312 19.44C18.9236 20.4843 18.185 22.7 16.4445 22.7H7.55562C5.81513 22.7 5.07654 20.4843 6.46895 19.44L7.65413 18.5511C8.15996 18.1717 8.77519 17.9667 9.40747 17.9667H11.3001V16.0005C8.57046 15.7432 6.27425 13.9965 5.23266 11.5818C5.16475 11.5719 5.09716 11.5518 5.03197 11.5207L2.964 10.5341C1.9474 10.0491 1.30005 9.023 1.30005 7.89663V6.44444C1.30005 4.83054 2.60837 3.52222 4.22227 3.52222H4.63338V2.58607ZM17.9667 8.66687C17.9667 11.9614 15.2966 14.6321 12.0023 14.6333H12C8.70475 14.6333 6.03338 11.962 6.03338 8.66665V2.69999H17.9667V2.70604V2.71633V2.72662V2.73693V2.74725V2.75758V2.76792V2.77827V2.78863V2.799V2.80939V2.81978V2.83019V2.8406V2.85103V2.86147V2.87192V2.88238V2.89285V2.90333V2.91382V2.92433V2.93484V2.94536V2.9559V2.96644V2.977V2.98757V2.99814V3.00873V3.01933V3.02994V3.04056V3.05119V3.06183V3.07247V3.08313V3.09381V3.10449V3.11518V3.12588V3.13659V3.14731V3.15804V3.16878V3.17953V3.19029V3.20106V3.21184V3.22264V3.23344V3.24425V3.25507V3.2659V3.27674V3.28759V3.29845V3.30931V3.32019V3.33108V3.34198V3.35289V3.3638V3.37473V3.38567V3.39661V3.40757V3.41853V3.4295V3.44049V3.45148V3.46248V3.47349V3.48451V3.49554V3.50658V3.51763V3.52868V3.53975V3.55082V3.56191V3.573V3.5841V3.59521V3.60633V3.61746V3.6286V3.63975V3.6509V3.66206V3.67324V3.68442V3.69561V3.70681V3.71801V3.72923V3.74045V3.75169V3.76293V3.77418V3.78544V3.7967V3.80798V3.81926V3.83056V3.84186V3.85317V3.86448V3.87581V3.88714V3.89848V3.90983V3.92119V3.93256V3.94393V3.95531V3.9667V3.9781V3.98951V4.00092V4.01234V4.02377V4.03521V4.04666V4.05811V4.06957V4.08104V4.09252V4.104V4.11549V4.12699V4.1385V4.15002V4.16154V4.17307V4.18461V4.19615V4.2077V4.21926V4.23083V4.2424V4.25398V4.26557V4.27717V4.28877V4.30038V4.312V4.32363V4.33526V4.3469V4.35854V4.3702V4.38186V4.39353V4.4052V4.41688V4.42857V4.44026V4.45197V4.46368V4.47539V4.48711V4.49884V4.51058V4.52232V4.53407V4.54582V4.55759V4.56936V4.58113V4.59291V4.6047V4.6165V4.6283V4.6401V4.65192V4.66374V4.67556V4.6874V4.69924V4.71108V4.72293V4.73479V4.74665V4.75852V4.7704V4.78228V4.79417V4.80607V4.81797V4.82987V4.84178V4.8537V4.86563V4.87755V4.88949V4.90143V4.91338V4.92533V4.93729V4.94925V4.96122V4.9732V4.98518V4.99717V5.00916V5.02115V5.03316V5.04517V5.05718V5.0692V5.08122V5.09325V5.10529V5.11733V5.12937V5.14142V5.15348V5.16554V5.17761V5.18968V5.20176V5.21384V5.22592V5.23802V5.25011V5.26221V5.27432V5.28643V5.29855V5.31067V5.32279V5.33492V5.34706V5.3592V5.37134V5.38349V5.39565V5.40781V5.41997V5.43214V5.44431V5.45649V5.46867V5.48085V5.49304V5.50524V5.51744V5.52964V5.54185V5.55406V5.56627V5.57849V5.59072V5.60295V5.61518V5.62742V5.63966V5.6519V5.66415V5.6764V5.68866V5.70092V5.71318V5.72545V5.73772V5.75V5.76228V5.77456V5.78685V5.79914V5.81144V5.82373V5.83604V5.84834V5.86065V5.87296V5.88528V5.8976V5.90992V5.92225V5.93458V5.94691V5.95925V5.97159V5.98393V5.99628V6.00863V6.02098V6.03334V6.0457V6.05806V6.07042V6.08279V6.09516V6.10754V6.11992V6.1323V6.14468V6.15707V6.16945V6.18185V6.19424V6.20664V6.21904V6.23144V6.24385V6.25626V6.26867V6.28108V6.2935V6.30591V6.31834V6.33076V6.34319V6.35561V6.36805V6.38048V6.39292V6.40535V6.41779V6.43024V6.44268V6.45513V6.46758V6.48003V6.49248V6.50494V6.5174V6.52986V6.54232V6.55478V6.56725V6.57972V6.59219V6.60466V6.61713V6.62961V6.64208V6.65456V6.66704V6.67953V6.69201V6.7045V6.71698V6.72947V6.74196V6.75445V6.76695V6.77944V6.79194V6.80444V6.81694V6.82944V6.84194V6.85444V6.86695V6.87945V6.89196V6.90447V6.91698V6.92949V6.942V6.95451V6.96703V6.97954V6.99206V7.00458V7.01709V7.02961V7.04213V7.05465V7.06717V7.0797V7.09222V7.10474V7.11727V7.12979V7.14232V7.15485V7.16737V7.1799V7.19243V7.20496V7.21749V7.23002V7.24255V7.25508V7.26761V7.28014V7.29268V7.30521V7.31774V7.33027V7.3428V7.35534V7.36787V7.38041V7.39294V7.40547V7.41801V7.43054V7.44307V7.45561V7.46814V7.48068V7.49321V7.50574V7.51828V7.53081V7.54334V7.55587V7.56841V7.58094V7.59347V7.606V7.61853V7.63106V7.64359V7.65612V7.66865V7.68118V7.69371V7.70624V7.71876V7.73129V7.74381V7.75634V7.76886V7.78139V7.79391V7.80643V7.81895V7.83147V7.84399V7.85651V7.86903V7.88154V7.89406V7.90657V7.91909V7.9316V7.94411V7.95662V7.96913V7.98163V7.99414V8.00665V8.01915V8.03165V8.04415V8.05665V8.06915V8.08165V8.09414V8.10664V8.11913V8.13162V8.14411V8.1566V8.16908V8.18157V8.19405V8.20653V8.21901V8.23149V8.24396V8.25644V8.26891V8.28138V8.29385V8.30631V8.31878V8.33124V8.3437V8.35616V8.36862V8.38107V8.39352V8.40597V8.41842V8.43087V8.44331V8.45575V8.46819V8.48063V8.49306V8.50549V8.51792V8.53035V8.54277V8.55519V8.56761V8.58003V8.59244V8.60486V8.61726V8.62967V8.64207V8.65448V8.66687ZM19.2766 9.82233L20.4333 9.27052C20.9628 9.01788 21.3001 8.48337 21.3001 7.89663V6.44444C21.3001 5.60374 20.6185 4.92222 19.7778 4.92222H19.3667V4.92533V4.93729V4.94925V4.96122V4.9732V4.98518V4.99717V5.00916V5.02115V5.03316V5.04517V5.05718V5.0692V5.08122V5.09325V5.10529V5.11733V5.12937V5.14142V5.15348V5.16554V5.17761V5.18968V5.20176V5.21384V5.22592V5.23802V5.25011V5.26221V5.27432V5.28643V5.29855V5.31067V5.32279V5.33492V5.34706V5.3592V5.37134V5.38349V5.39565V5.40781V5.41997V5.43214V5.44431V5.45649V5.46867V5.48085V5.49304V5.50524V5.51744V5.52964V5.54185V5.55406V5.56627V5.57849V5.59072V5.60295V5.61518V5.62742V5.63966V5.6519V5.66415V5.6764V5.68866V5.70092V5.71318V5.72545V5.73772V5.75V5.76228V5.77456V5.78685V5.79914V5.81144V5.82373V5.83604V5.84834V5.86065V5.87296V5.88528V5.8976V5.90992V5.92225V5.93458V5.94691V5.95925V5.97159V5.98393V5.99628V6.00863V6.02098V6.03334V6.0457V6.05806V6.07042V6.08279V6.09516V6.10754V6.11992V6.1323V6.14468V6.15707V6.16945V6.18185V6.19424V6.20664V6.21904V6.23144V6.24385V6.25626V6.26867V6.28108V6.2935V6.30591V6.31834V6.33076V6.34319V6.35561V6.36805V6.38048V6.39292V6.40535V6.41779V6.43024V6.44268V6.45513V6.46758V6.48003V6.49248V6.50494V6.5174V6.52986V6.54232V6.55478V6.56725V6.57972V6.59219V6.60466V6.61713V6.62961V6.64208V6.65456V6.66704V6.67953V6.69201V6.7045V6.71698V6.72947V6.74196V6.75445V6.76695V6.77944V6.79194V6.80444V6.81694V6.82944V6.84194V6.85444V6.86695V6.87945V6.89196V6.90447V6.91698V6.92949V6.942V6.95451V6.96703V6.97954V6.99206V7.00458V7.01709V7.02961V7.04213V7.05465V7.06717V7.0797V7.09222V7.10474V7.11727V7.12979V7.14232V7.15485V7.16737V7.1799V7.19243V7.20496V7.21749V7.23002V7.24255V7.25508V7.26761V7.28014V7.29268V7.30521V7.31774V7.33027V7.3428V7.35534V7.36787V7.38041V7.39294V7.40547V7.41801V7.43054V7.44307V7.45561V7.46814V7.48068V7.49321V7.50574V7.51828V7.53081V7.54334V7.55587V7.56841V7.58094V7.59347V7.606V7.61853V7.63106V7.64359V7.65612V7.66865V7.68118V7.69371V7.70624V7.71876V7.73129V7.74381V7.75634V7.76886V7.78139V7.79391V7.80643V7.81895V7.83147V7.84399V7.85651V7.86903V7.88154V7.89406V7.90657V7.91909V7.9316V7.94411V7.95662V7.96913V7.98163V7.99414V8.00665V8.01915V8.03165V8.04415V8.05665V8.06915V8.08165V8.09414V8.10664V8.11913V8.13162V8.14411V8.1566V8.16908V8.18157V8.19405V8.20653V8.21901V8.23149V8.24396V8.25644V8.26891V8.28138V8.29385V8.30631V8.31878V8.33124V8.3437V8.35616V8.36862V8.38107V8.39352V8.40597V8.41842V8.43087V8.44331V8.45575V8.46819V8.48063V8.49306V8.50549V8.51792V8.53035V8.54277V8.55519V8.56761V8.58003V8.59244V8.60486V8.61726V8.62967V8.64207V8.65448V8.66687C19.3667 9.05998 19.3359 9.44589 19.2766 9.82233ZM4.63338 8.66665C4.63338 9.05984 4.66419 9.44583 4.72351 9.82235L3.56682 9.27052C3.03727 9.01788 2.70005 8.48337 2.70005 7.89663V6.44444C2.70005 5.60374 3.38157 4.92222 4.22227 4.92222H4.63338V8.66665ZM8.49413 19.6711C8.75763 19.4735 9.0781 19.3667 9.40747 19.3667H12.0001H14.5927C14.922 19.3667 15.2425 19.4735 15.506 19.6711L16.6912 20.56C17.0072 20.797 16.8396 21.3 16.4445 21.3H7.55562C7.16053 21.3 6.99289 20.797 7.30895 20.56L8.49413 19.6711Z"
+//                       fill="white"
+//                     />
+//                     <path
+//                       d="M11.8425 6.20125C11.9226 6.09893 12.0774 6.09893 12.1575 6.20125L12.7183 6.91782C12.7413 6.94724 12.7721 6.96963 12.8072 6.98244L13.662 7.29436C13.7841 7.3389 13.832 7.48617 13.7594 7.59395L13.2512 8.34874C13.2303 8.37973 13.2185 8.41595 13.2172 8.45329L13.1847 9.36265C13.1801 9.49249 13.0548 9.58351 12.9299 9.5478L12.055 9.29771C12.019 9.28744 11.981 9.28744 11.945 9.29771L11.0701 9.5478C10.9452 9.58351 10.8199 9.49249 10.8153 9.36265L10.7828 8.45329C10.7815 8.41595 10.7697 8.37973 10.7488 8.34874L10.2406 7.59395C10.168 7.48617 10.2159 7.3389 10.338 7.29436L11.1928 6.98244C11.2279 6.96963 11.2587 6.94724 11.2817 6.91782L11.8425 6.20125Z"
+//                       fill="#DF1463"
+//                     />
+//                   </svg>
+//                   <div className="flex items-center gap-1 rounded-full bg-green-400/25 py-0.5 pl-1 pr-2">
+//                     <p className="-rotate-90 leading-tight text-green-500">
+//                       &rarr;
+//                     </p>
+//                     <p className="text-xs leading-tight text-white">4%</p>
+//                   </div>
+//                   <p className="text-white opacity-0 delay-200 duration-700 ease-in-out group-hover:opacity-100">
+//                     75 Orders
+//                   </p>
+//                 </div>
+//                 {/* item 2 */}
+//                 <div className="flex items-center gap-2">
+//                   <svg
+//                     xmlns="http://www.w3.org/2000/svg"
+//                     width="20px"
+//                     height="20px"
+//                     viewBox="0 0 24 24"
+//                     fill="none"
+//                   >
+//                     <path
+//                       d="M3.7644 13.3726V11.0849C3.7644 6.53663 7.45154 2.84949 11.9999 2.84949V2.84949C16.5482 2.84949 20.2353 6.53663 20.2353 11.0849V13.9856"
+//                       stroke="#DF1463"
+//                       strokeWidth="1.69904"
+//                     />
+//                     <path
+//                       d="M2.84937 17.4903V16.1566C2.84937 14.1351 4.4881 12.4964 6.50957 12.4964C7.5203 12.4964 8.33967 13.3158 8.33967 14.3265V19.3204C8.33967 20.3311 7.5203 21.1505 6.50957 21.1505C4.4881 21.1505 2.84937 19.5118 2.84937 17.4903Z"
+//                       stroke="white"
+//                       strokeWidth="1.69904"
+//                     />
+//                     <path
+//                       d="M21.1504 17.4903V16.1566C21.1504 14.1351 19.5117 12.4964 17.4902 12.4964C16.4795 12.4964 15.6601 13.3158 15.6601 14.3265V19.3204C15.6601 20.3311 16.4795 21.1505 17.4902 21.1505C19.5117 21.1505 21.1504 19.5118 21.1504 17.4903Z"
+//                       stroke="white"
+//                       strokeWidth="1.69904"
+//                     />
+//                   </svg>
+//                   <div className="flex items-center gap-1 rounded-full bg-green-400/25 py-0.5 pl-1 pr-2">
+//                     <p className="-rotate-90 leading-tight text-green-500">
+//                       &rarr;
+//                     </p>
+//                     <p className="text-xs leading-tight text-white">4%</p>
+//                   </div>
+//                   <p className="text-white opacity-0 delay-200 duration-700 ease-in-out group-hover:opacity-100">
+//                     357 Delivered
+//                   </p>
+//                 </div>
+//                 {/* item 3 */}
+//                 <div className="flex items-center gap-2">
+//                   <svg
+//                     xmlns="http://www.w3.org/2000/svg"
+//                     width="20px"
+//                     height="20px"
+//                     viewBox="0 0 24 24"
+//                     fill="none"
+//                   >
+//                     <path
+//                       d="M3.7644 13.3726V11.0849C3.7644 6.53663 7.45154 2.84949 11.9999 2.84949V2.84949C16.5482 2.84949 20.2353 6.53663 20.2353 11.0849V13.9856"
+//                       stroke="#DF1463"
+//                       strokeWidth="1.69904"
+//                     />
+//                     <path
+//                       d="M2.84937 17.4903V16.1566C2.84937 14.1351 4.4881 12.4964 6.50957 12.4964C7.5203 12.4964 8.33967 13.3158 8.33967 14.3265V19.3204C8.33967 20.3311 7.5203 21.1505 6.50957 21.1505C4.4881 21.1505 2.84937 19.5118 2.84937 17.4903Z"
+//                       stroke="white"
+//                       strokeWidth="1.69904"
+//                     />
+//                     <path
+//                       d="M21.1504 17.4903V16.1566C21.1504 14.1351 19.5117 12.4964 17.4902 12.4964C16.4795 12.4964 15.6601 13.3158 15.6601 14.3265V19.3204C15.6601 20.3311 16.4795 21.1505 17.4902 21.1505C19.5117 21.1505 21.1504 19.5118 21.1504 17.4903Z"
+//                       stroke="white"
+//                       strokeWidth="1.69904"
+//                     />
+//                   </svg>
+//                   <div className="flex items-center gap-1 rounded-full bg-red-400/25 py-0.5 pl-1 pr-2">
+//                     <p className="-rotate-n90 leading-tight text-red-500">
+//                       &rarr;
+//                     </p>
+//                     <p className="text-xs leading-tight text-white">25%</p>
+//                   </div>
+//                   <p className="text-white opacity-0 delay-200 duration-700 ease-in-out group-hover:opacity-100">
+//                     65 Canceled
+//                   </p>
+//                 </div>
+//               </div>
+//             </div>
+//             {/* main */}
+//             <div
+//               className="absolute -right-16 top-0 z-20 flex w-96 flex-col gap-4 self-end rounded-xl rounded-l-2xl border-none px-5 py-3 text-base font-semibold leading-7 transition-all duration-700 group-hover:-right-14 group-hover:w-64 group-hover:rounded-l-lg"
+//               style={{ backgroundColor: "#B0C4DE" }}
+//             >
+//               <p className="text-[#fff]">Total Orders</p>
+//               <p className="text-[#fff]">Total Delivered</p>
+//               <p className="text-[#fff]">Total Canceled</p>
+//             </div>
+//             <div
+//               className="h-16 w-[28rem] -left-10 shadow-2xl shadow-[#3d348b] absolute bottom-0"
+//               style={{ backgroundColor: "#99CCFF" }}
+//             ></div>
+//           </div>
+//         </div>
+//         <OrderTable />
+//       </div>
+//     </div>
+//   );
+// };
 
-export default OrderComponent;
+// export default OrderComponent;
