+++ src/components/layouts/order/index.tsx
@@ -1,16 +1,16 @@
-import { AppProvider } from "@/contexts/AppContext"
-import { AuthProvider } from "@/contexts/JWTContext"
-import OrderComponent from "./OrderComponent"
+// import { AppProvider } from "@/contexts/AppContext"
+// import { AuthProvider } from "@/contexts/JWTContext"
+// import OrderComponent from "./OrderComponent"
 
 
-const Order = (props : {}) => {
-    return (
-        <AppProvider>
-            <AuthProvider>
-                <OrderComponent/>
-            </AuthProvider>
-        </AppProvider>
-    )
-}
+// const Order = (props : {}) => {
+//     return (
+//         <AppProvider>
+//             <AuthProvider>
+//                 <OrderComponent/>
+//             </AuthProvider>
+//         </AppProvider>
+//     )
+// }
 
-export default Order
+// export default Order
