+++ src/components/layouts/sidebarComponent/SideBarComponent.tsx
@@ -1,3 +1,4 @@
+"use client"
 import path from "path";
 import "./SideBarComponent.css";
 import { title } from "process";
@@ -45,7 +46,7 @@ const SideBarComponent = (props: {}) => {
         <div className="flex h-screen flex-col justify-between pt-2 pb-6">
           <div>
             <div
-            className={`w-max p-2.5 ${isHovered ? "block" : "hidden"}`}
+              className={`w-max p-2.5 ${isHovered ? "block" : "hidden"}`}
             >
               <img
                 src="/Icons/logo.png"
@@ -56,46 +57,55 @@ const SideBarComponent = (props: {}) => {
             </div>
             <ul className="mt-6 space-y-2 tracking-wide">
               <li className="min-w-max">
-                <a
-                  href="#"
-                  aria-label="dashboard"
-                  className="relative flex items-center space-x-4 bg-gradient-to-r from-sky-600 to-cyan-400 px-4 py-3 text-white"
+                <div onClick={() => {
+                  navigateTo(PATH_MAIN.root)
+                }}
+                className={`links_hover ${typeof window !== "undefined" &&
+                    isActive(PATH_MAIN.root)
+                    ? "active_current_link"
+                    : ""
+                    }`}
                 >
-                  <svg
-                    className="-ml-1 h-6 w-6"
-                    viewBox="0 0 24 24"
-                    fill="none"
+                  <a
+                    href="/"
+                    aria-label="dashboard"
+                    className="relative flex items-center rounded-full space-x-4 text-gray-700 hover:bg-gradient-to-r hover:from-sky-600 hover:to-cyan-400 px-4 py-3 hover:text-white"
                   >
-                    <path
-                      d="M6 8a2 2 0 0 1 2-2h1a2 2 0 0 1 2 2v1a2 2 0 0 1-2 2H8a2 2 0 0 1-2-2V8ZM6 15a2 2 0 0 1 2-2h1a2 2 0 0 1 2 2v1a2 2 0 0 1-2 2H8a2 2 0 0 1-2-2v-1Z"
-                      className="fill-current text-cyan-400 dark:fill-slate-600"
-                    ></path>
-                    <path
-                      d="M13 8a2 2 0 0 1 2-2h1a2 2 0 0 1 2 2v1a2 2 0 0 1-2 2h-1a2 2 0 0 1-2-2V8Z"
-                      className="fill-current text-cyan-200 group-hover:text-cyan-300"
-                    ></path>
-                    <path
-                      d="M13 15a2 2 0 0 1 2-2h1a2 2 0 0 1 2 2v1a2 2 0 0 1-2 2h-1a2 2 0 0 1-2-2v-1Z"
-                      className="fill-current group-hover:text-sky-300"
-                    ></path>
-                  </svg>
-                  <span className="-mr-1 font-medium">Dashboard</span>
-                </a>
+                    <svg
+                      className="-ml-1 h-6 w-6"
+                      viewBox="0 0 24 24"
+                      fill="none"
+                    >
+                      <path
+                        d="M6 8a2 2 0 0 1 2-2h1a2 2 0 0 1 2 2v1a2 2 0 0 1-2 2H8a2 2 0 0 1-2-2V8ZM6 15a2 2 0 0 1 2-2h1a2 2 0 0 1 2 2v1a2 2 0 0 1-2 2H8a2 2 0 0 1-2-2v-1Z"
+                        className="fill-current text-cyan-400 dark:fill-slate-600"
+                      ></path>
+                      <path
+                        d="M13 8a2 2 0 0 1 2-2h1a2 2 0 0 1 2 2v1a2 2 0 0 1-2 2h-1a2 2 0 0 1-2-2V8Z"
+                        className="fill-current text-cyan-200 group-hover:text-cyan-300"
+                      ></path>
+                      <path
+                        d="M13 15a2 2 0 0 1 2-2h1a2 2 0 0 1 2 2v1a2 2 0 0 1-2 2h-1a2 2 0 0 1-2-2v-1Z"
+                        className="fill-current group-hover:text-sky-300"
+                      ></path>
+                    </svg>
+                    <span className="-mr-1 font-medium">Dashboard</span>
+                  </a>
+                </div>
               </li>
               <li className="min-w-max">
                 <div onClick={() => {
                   navigateTo(PATH_MAIN.category)
                 }}
-                className={`links_hover relative font-medium inline-block whitespace-nowrap ${
-                  typeof window !== "undefined" &&
-                  isActive(PATH_MAIN.category)
+                  className={`links_hover ${typeof window !== "undefined" &&
+                    isActive(PATH_MAIN.category)
                     ? "active_current_link"
                     : ""
-                }`}
+                    }`}
                 >
                   <a
                     href="/category"
-                    className="bg group flex items-center space-x-4 rounded-full px-4 py-3 text-gray-600"
+                    className="relative flex items-center rounded-full space-x-4 text-gray-700 hover:bg-gradient-to-r hover:from-sky-600 hover:to-cyan-400 px-4 py-3 hover:text-white"
                   >
                     <svg
                       xmlns="http://www.w3.org/2000/svg"
@@ -119,34 +129,43 @@ const SideBarComponent = (props: {}) => {
                 </div>
               </li>
               <li className="min-w-max">
-                <a
-                  href="#"
-                  className="group flex items-center space-x-4 rounded-md px-4 py-3 text-gray-600"
-                >
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    className="h-5 w-5"
-                    viewBox="0 0 20 20"
-                    fill="currentColor"
+                <div onClick={() => {
+                  navigateTo(PATH_MAIN.order)
+                }}
+                  className={`links_hover ${typeof window !== "undefined" &&
+                      isActive(PATH_MAIN.order)
+                      ? "active_current_link"
+                      : ""
+                    }`}>
+                  <a
+                    href="/order"
+                    className="relative flex items-center rounded-full space-x-4 text-gray-700 hover:bg-gradient-to-r hover:from-sky-600 hover:to-cyan-400 px-4 py-3 hover:text-white"
                   >
-                    <path
-                      className="fill-current text-gray-600 group-hover:text-cyan-600"
-                      fill-rule="evenodd"
-                      d="M2 5a2 2 0 012-2h8a2 2 0 012 2v10a2 2 0 002 2H4a2 2 0 01-2-2V5zm3 1h6v4H5V6zm6 6H5v2h6v-2z"
-                      clip-rule="evenodd"
-                    />
-                    <path
-                      className="fill-current text-gray-300 group-hover:text-cyan-300"
-                      d="M15 7h1a2 2 0 012 2v5.5a1.5 1.5 0 01-3 0V7z"
-                    />
-                  </svg>
-                  <span className="group-hover:text-gray-700">Order</span>
-                </a>
+                    <svg
+                      xmlns="http://www.w3.org/2000/svg"
+                      className="h-5 w-5"
+                      viewBox="0 0 20 20"
+                      fill="currentColor"
+                    >
+                      <path
+                        className="fill-current text-gray-600 group-hover:text-cyan-600"
+                        fill-rule="evenodd"
+                        d="M2 5a2 2 0 012-2h8a2 2 0 012 2v10a2 2 0 002 2H4a2 2 0 01-2-2V5zm3 1h6v4H5V6zm6 6H5v2h6v-2z"
+                        clip-rule="evenodd"
+                      />
+                      <path
+                        className="fill-current text-gray-300 group-hover:text-cyan-300"
+                        d="M15 7h1a2 2 0 012 2v5.5a1.5 1.5 0 01-3 0V7z"
+                      />
+                    </svg>
+                    <span className="group-hover:text-gray-700">Order</span>
+                  </a>
+                </div>
               </li>
               <li className="min-w-max">
                 <a
                   href="#"
-                  className="group flex items-center space-x-4 rounded-md px-4 py-3 text-gray-600"
+                  className="relative flex items-center rounded-full space-x-4 text-gray-700 hover:bg-gradient-to-r hover:from-sky-600 hover:to-cyan-400 px-4 py-3 hover:text-white"
                 >
                   <svg
                     xmlns="http://www.w3.org/2000/svg"
@@ -169,7 +188,7 @@ const SideBarComponent = (props: {}) => {
               <li className="min-w-max">
                 <a
                   href="#"
-                  className="group flex items-center space-x-4 rounded-md px-4 py-3 text-gray-600"
+                  className="relative flex items-center rounded-full space-x-4 text-gray-700 hover:bg-gradient-to-r hover:from-sky-600 hover:to-cyan-400 px-4 py-3 hover:text-white"
                 >
                   <svg
                     xmlns="http://www.w3.org/2000/svg"
@@ -194,7 +213,7 @@ const SideBarComponent = (props: {}) => {
               <li className="min-w-max">
                 <a
                   href="#"
-                  className="group flex items-center space-x-4 rounded-md px-4 py-3 text-gray-600"
+                  className="relative flex items-center rounded-full space-x-4 text-gray-700 hover:bg-gradient-to-r hover:from-sky-600 hover:to-cyan-400 px-4 py-3 hover:text-white"
                 >
                   <svg
                     xmlns="http://www.w3.org/2000/svg"
@@ -221,7 +240,7 @@ const SideBarComponent = (props: {}) => {
           <div className="w-max -mb-3">
             <a
               href="#"
-              className="group flex items-center space-x-4 rounded-md px-4 py-3 text-gray-600"
+              className="relative flex items-center rounded-full space-x-4 text-gray-700 hover:bg-gradient-to-r hover:from-sky-600 hover:to-cyan-400 px-4 py-3 hover:text-white"
             >
               <svg
                 xmlns="http://www.w3.org/2000/svg"
