+++ src/contexts/AppContext.tsx
@@ -1,3 +1,4 @@
+"use client"
 import { ReactNode, createContext, useEffect, useReducer } from "react";
 
 type AppContextState = {
