+++ src/contexts/AuthGoogleContext.tsx
@@ -1,3 +1,4 @@
+"use client"
 import { useContext, createContext, useEffect, useState } from "react";
 import {
   GoogleAuthProvider,
